Add support for component metadata suppliers for Maven repositories
This implements feature parity between Ivy and Maven repositories with regards to
component metadata suppliers. Before, only Ivy repositories could benefit from
user provided component metadata suppliers. Now, both Maven and Ivy repositories
support this. The test that was Ivy specific has been migrated to the new test
fixtures, allowing us to check both Maven and Ivy in a single test class.
