+++ subprojects/dependency-management/src/main/java/org/gradle/api/internal/artifacts/ivyservice/ivyresolve/CachingModuleComponentRepository.java
@@ -18,7 +18,6 @@ package org.gradle.api.internal.artifacts.ivyservice.ivyresolve;
 import org.gradle.api.Transformer;
 import org.gradle.api.artifacts.ArtifactIdentifier;
 import org.gradle.api.artifacts.ComponentMetadataSupplier;
-import org.gradle.api.artifacts.ComponentMetadataVersionLister;
 import org.gradle.api.artifacts.ModuleIdentifier;
 import org.gradle.api.artifacts.ModuleVersionIdentifier;
 import org.gradle.api.artifacts.component.ComponentArtifactIdentifier;
@@ -129,11 +128,6 @@ public class CachingModuleComponentRepository implements ModuleComponentReposito
 
     public ComponentMetadataSupplier createMetadataSupplier() {
         return delegate.createMetadataSupplier();
-    }
-
-    @Override
-    public ComponentMetadataVersionLister createVersionLister() {
-        return delegate.createVersionLister();
     }
 
     @Override
