+++ subprojects/dependency-management/src/testFixtures/groovy/org/gradle/integtests/resolve/AbstractModuleDependencyResolveTest.groovy
@@ -123,9 +123,21 @@ abstract class AbstractModuleDependencyResolveTest extends AbstractHttpDependenc
         """
     }
 
+    void setMetadataSupplierClassWithParams(String clazz, String... params) {
+        buildFile << """
+            repositories."${useIvy()?'ivy':'maven'}".setMetadataSupplier($clazz) { ${params.collect { "params($it)" }.join(';')} }
+        """
+    }
+
     void setMetadataListerClass(String clazz) {
         buildFile << """
             repositories."${useIvy()?'ivy':'maven'}".componentVersionsLister = $clazz
+        """
+    }
+
+    void setMetadataListerClassWithParams(String clazz, String... params) {
+        buildFile << """
+            repositories."${useIvy()?'ivy':'maven'}".setComponentVersionsLister($clazz) { ${params.collect { "params($it)" }.join(';')} }
         """
     }
