+++ subprojects/internal-integ-testing/src/main/groovy/org/gradle/integtests/fixtures/GradleMetadataResolveRunner.groovy
@@ -48,4 +48,8 @@ class GradleMetadataResolveRunner extends BehindFlagFeatureRunner {
     static boolean useIvy() {
         System.getProperty(REPOSITORY_TYPE) == "ivy"
     }
+
+    static boolean useMaven() {
+        System.getProperty(REPOSITORY_TYPE) == "maven"
+    }
 }
